Moving a couple shard methods to topo.
From topotools, they really belong in topo. The test has to stay in
topotools, as it depends on fake topo implementation, that depends on
topo.
